v3.1
Changed the colour that the buttuns highlight into when they get
hovered.
Added a new class with appropriate fields in the cunstructor and made an
ArrayList for that class. This will be used to add and retrieve the
weapons that will be displayed for the user to see.
Totally changed the design of the second logo. New one to come soon.